add: [core] Seperation of test database conf from test instance launcher
