+++ common_tests/test_test_database.py
@@ -6,6 +6,7 @@ import sys
 
 sys.path.append(os.path.abspath(os.path.pardir))
 import common.TestDBHandler.test_instance_launcher as test_database_handler
+import common.TestDBHandler.test_database_only_conf as test_database_only_conf
 from common.environment_variable import get_homedir
 
 import pathlib
@@ -22,7 +23,7 @@ class test_tests_database_launcher(unittest.TestCase):
         self.test_file_path = pathlib.Path.cwd() / pathlib.Path("tests/test_files")
 
         self.test_db_handler = test_database_handler.TestInstanceLauncher()
-        db_test_conf = test_database_handler.TestInstance_database_conf()
+        db_test_conf = test_database_only_conf.TestInstance_database_conf()
         print("\n[TESTS] LAUNCHING DATABASE AS TEST : NOTHING WILL BE WRITEN ON STORAGE OR CACHE DATABASES [TESTS]\n")
         self.test_db_handler.create_full_instance(db_conf=db_test_conf)
